pyproject.toml
[project]
name = "uv1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-openai",
    "langgraph",
    "python-dotenv",
    "flask",
    "transformers",
    "torch",
    "langchain-huggingface",
]
